Game setup
Addition of consumables and their effects, as well as equipment and their effects.
Fixed the imported Dungeon Generator, so that it always works first try. Also removed a while loop, making it a bit cleaner.
Added new State Machine (credit to Valentijn Muijrers) and compressed the Node class into the AStar class